arregle bugs, corregi la obtencion de la obra, arregle el registro de la obra, del capitulo, faltan rutas y cosas por corregir

+++ routes/Capitulo.py
@@ -1,7 +1,10 @@
-from fastapi import APIRouter, UploadFile, Path, Body
+from fastapi import APIRouter, UploadFile, Path, Body, HTTPException
 from models.CapituloModel import CapituloModel, StrategyCapituloArts, StrategyCapituloScan
 from models.entities.Capitulo import Capitulo
+from helper.sb_help import get_role_user, user_exist
 from uuid import UUID
+from models.ObraModel import ObraModel
+from typing import List
 
 mainCapi = APIRouter(prefix="/cap", tags=['capitulo'], responses={404: {"description": "Not found"}})
 
@@ -14,10 +17,16 @@ async def get_capi(obra_id: UUID = Path(...), numero: int = Path(...)):
         return ({'message':str(ex)}),500
     
 @mainCapi.post('/addcapi')
-async def add_capitulo(numero: int = Body(...), fecha: str = Body(...), idioma: str = Body(...), obra: UUID = Body(...)):
+async def add_capitulo(numero: int = Body(...), fecha: str = Body(...), idioma: str = Body(...), obra: UUID = Body(0), price: float = Body(...), images: List[str] = Body(...)):
     try:
-        capi = Capitulo('', numero, fecha, idioma)
-        affec_row = CapituloModel.add_capitulo(capi, obra)
+        obram = ObraModel()
+        if not obram.exist_obra(obra) and obram.oneshot_bool(obra):
+            raise HTTPException(status_code=404, detail="No se puede agregar el capitulo, perdon uwu")
+        capim = CapituloModel()
+        if capim.get_max_capitulo(obra) >= numero:
+            raise HTTPException(status_code=404, detail="El capitulo ya existe")
+        capi = Capitulo(numero, fecha, idioma,price, images)
+        affec_row = StrategyCapituloArts.add_capitulo(capi, obra)
         if affec_row == 0:
             return ({'message': "Error on insert"})
         return ({'message':"Ok"})
@@ -26,13 +35,27 @@ async def add_capitulo(numero: int = Body(...), fecha: str = Body(...), idioma: 
 @mainCapi.delete('/delete/{numero}/{id_obra}/{id_user}')
 async def delete_capitulo(numero: int = Path(...), id_obra: UUID = Path(...), id_user: UUID = Path(...)):
     try:
-        affec_row = CapituloModel(StrategyCapituloArts).delete_capi(numero, id_obra, id_user)
+        if not user_exist(id_user):
+            raise HTTPException(status_code=404, detail="User not exist")
+        if get_role_user(id_user) == "artist":
+            affec_row = CapituloModel(StrategyCapituloArts).delete_capi(numero, id_obra, id_user)
+        elif get_role_user(id_user) == "scan":
+            affec_row = CapituloModel(StrategyCapituloScan).delete_capi(numero, id_obra, id_user)
+        else:
+            raise HTTPException(status_code=403, detail="No permitido")
         if affec_row == 0:
             return ({'message': "Error on delete"})
         return ({'message':"Ok"})
     except Exception as ex:
         return ({'message':str(ex)}),500
+@mainCapi.put('/update/{id_obra}/{numero}')
+async def update_capitulo(numero: int = Path(...), id_obra: UUID = Path(...), oldpages: List[str] = Body(...), newpages: List[str] = Body(...), orden: List[int] = Body(...), id_user: UUID = Body(...)):
+    try:
+       pass
+    except Exception as ex:
+        return ({'message':str(ex)}),500
 
+"""
 @mainCapi.delete('/deletescan/{numero}/{id_obra}/{id_scan}')
 async def delete_capitulo(numero: int = Path(...), id_obra: UUID = Path(...), id_scan: UUID = Path(...)):
     try:
@@ -41,4 +64,4 @@ async def delete_capitulo(numero: int = Path(...), id_obra: UUID = Path(...), id
             return ({'message': "Error on delete"})
         return ({'message':"Ok"})
     except Exception as ex:
-        return ({'message':str(ex)}),500
+        return ({'message':str(ex)}),500"""
